application/Frontend/requirements.txt
streamlit
requests
python-docx
pymupdf
reportlab
googletrans-py
PyPDF2
openpyxl